refactor: split out `model_manager` module
