+++ src/speaches/model_manager.py
@@ -1,34 +1,10 @@
-from __future__ import annotations
-
-from collections import OrderedDict
+from collections.abc import Callable
 import gc
-import json
 import logging
-from pathlib import Path
 import threading
 import time
-from typing import TYPE_CHECKING
-
-from faster_whisper import WhisperModel
-from kokoro_onnx import Kokoro
-from onnxruntime import InferenceSession
-
-from speaches.executors.kokoro.utils import get_kokoro_model_path
-from speaches.executors.piper.utils import get_piper_voice_model_file
-
-if TYPE_CHECKING:
-    from collections.abc import Callable
-
-    from piper.voice import PiperVoice
-
-    from speaches.config import (
-        WhisperConfig,
-    )
 
 logger = logging.getLogger(__name__)
-
-
-# TODO: enable concurrent model downloads
 
 
 class SelfDisposingModel[T]:
@@ -104,133 +80,3 @@ class SelfDisposingModel[T]:
 
     def __exit__(self, *_args) -> None:  # noqa: ANN002
         self._decrement_ref()
-
-
-class WhisperModelManager:
-    def __init__(self, whisper_config: WhisperConfig) -> None:
-        self.whisper_config = whisper_config
-        self.loaded_models: OrderedDict[str, SelfDisposingModel[WhisperModel]] = OrderedDict()
-        self._lock = threading.Lock()
-
-    def _load_fn(self, model_id: str) -> WhisperModel:
-        return WhisperModel(
-            model_id,
-            device=self.whisper_config.inference_device,
-            device_index=self.whisper_config.device_index,
-            compute_type=self.whisper_config.compute_type,
-            cpu_threads=self.whisper_config.cpu_threads,
-            num_workers=self.whisper_config.num_workers,
-        )
-
-    def _handle_model_unloaded(self, model_id: str) -> None:
-        with self._lock:
-            if model_id in self.loaded_models:
-                del self.loaded_models[model_id]
-
-    def unload_model(self, model_id: str) -> None:
-        with self._lock:
-            model = self.loaded_models.get(model_id)
-            if model is None:
-                raise KeyError(f"Model {model_id} not found")
-            # WARN: ~300 MB of memory will still be held by the model. See https://github.com/SYSTRAN/faster-whisper/issues/992
-            self.loaded_models[model_id].unload()
-
-    def load_model(self, model_id: str) -> SelfDisposingModel[WhisperModel]:
-        logger.debug(f"Loading model {model_id}")
-        with self._lock:
-            logger.debug("Acquired lock")
-            if model_id in self.loaded_models:
-                logger.debug(f"{model_id} model already loaded")
-                return self.loaded_models[model_id]
-            self.loaded_models[model_id] = SelfDisposingModel[WhisperModel](
-                model_id,
-                load_fn=lambda: self._load_fn(model_id),
-                ttl=self.whisper_config.ttl,
-                model_unloaded_callback=self._handle_model_unloaded,
-            )
-            return self.loaded_models[model_id]
-
-
-ONNX_PROVIDERS = ["CUDAExecutionProvider", "CPUExecutionProvider"]
-
-
-class PiperModelManager:
-    def __init__(self, ttl: int) -> None:
-        self.ttl = ttl
-        self.loaded_models: OrderedDict[str, SelfDisposingModel[PiperVoice]] = OrderedDict()
-        self._lock = threading.Lock()
-
-    def _load_fn(self, model_id: str) -> PiperVoice:
-        from piper.voice import PiperConfig, PiperVoice
-
-        model_path = get_piper_voice_model_file(model_id)
-        inf_sess = InferenceSession(model_path, providers=ONNX_PROVIDERS)
-        config_path = Path(str(model_path) + ".json")
-        conf = PiperConfig.from_dict(json.loads(config_path.read_text()))
-        return PiperVoice(session=inf_sess, config=conf)
-
-    def _handle_model_unloaded(self, model_id: str) -> None:
-        with self._lock:
-            if model_id in self.loaded_models:
-                del self.loaded_models[model_id]
-
-    def unload_model(self, model_id: str) -> None:
-        with self._lock:
-            model = self.loaded_models.get(model_id)
-            if model is None:
-                raise KeyError(f"Model {model_id} not found")
-            self.loaded_models[model_id].unload()
-
-    def load_model(self, model_id: str) -> SelfDisposingModel[PiperVoice]:
-        from piper.voice import PiperVoice
-
-        with self._lock:
-            if model_id in self.loaded_models:
-                logger.debug(f"{model_id} model already loaded")
-                return self.loaded_models[model_id]
-            self.loaded_models[model_id] = SelfDisposingModel[PiperVoice](
-                model_id,
-                load_fn=lambda: self._load_fn(model_id),
-                ttl=self.ttl,
-                model_unloaded_callback=self._handle_model_unloaded,
-            )
-            return self.loaded_models[model_id]
-
-
-class KokoroModelManager:
-    def __init__(self, ttl: int) -> None:
-        self.ttl = ttl
-        self.loaded_models: OrderedDict[str, SelfDisposingModel[Kokoro]] = OrderedDict()
-        self._lock = threading.Lock()
-
-    # TODO
-    def _load_fn(self, _model_id: str) -> Kokoro:
-        model_path = get_kokoro_model_path()
-        voices_path = model_path.parent / "voices.bin"
-        inf_sess = InferenceSession(model_path, providers=ONNX_PROVIDERS)
-        return Kokoro.from_session(inf_sess, str(voices_path))
-
-    def _handle_model_unloaded(self, model_id: str) -> None:
-        with self._lock:
-            if model_id in self.loaded_models:
-                del self.loaded_models[model_id]
-
-    def unload_model(self, model_id: str) -> None:
-        with self._lock:
-            model = self.loaded_models.get(model_id)
-            if model is None:
-                raise KeyError(f"Model {model_id} not found")
-            self.loaded_models[model_id].unload()
-
-    def load_model(self, model_id: str) -> SelfDisposingModel[Kokoro]:
-        with self._lock:
-            if model_id in self.loaded_models:
-                logger.debug(f"{model_id} model already loaded")
-                return self.loaded_models[model_id]
-            self.loaded_models[model_id] = SelfDisposingModel[Kokoro](
-                model_id,
-                load_fn=lambda: self._load_fn(model_id),
-                ttl=self.ttl,
-                model_unloaded_callback=self._handle_model_unloaded,
-            )
-            return self.loaded_models[model_id]
